Consolidate compat event tests into one file
